pyproject.toml
[project]
name = "vacation-price-tracker"
version = "0.1.0"
description = "Vacation Price Tracker - Full Stack Application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "sqlmodel",
    "asyncpg",
    "alembic",
    "temporalio",
    "authlib",
    "httpx",
    "python-jose[cryptography]",
    "uvicorn",
    "redis>=5.0.0",
    "pydantic-settings",
    "greenlet",
    "itsdangerous",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "aiosqlite",
    "ruff",
    "pip-audit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# E: pycodestyle errors
# F: pyflakes
# I: isort (import sorting)
# W: pycodestyle warnings
# UP: pyupgrade (modern Python syntax)
# B: flake8-bugbear (common bugs)
# C90: mccabe (complexity)
select = ["E", "F", "I", "W", "UP", "B", "C90"]

# Ignore specific rules that are too strict
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = ["E712"]  # Allow `== None` in tests for clarity

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = ["apps/api/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["apps/api/tests", "apps/worker/tests"]

[tool.coverage.run]
source = ["apps/api/app"]
omit = [
    "apps/api/app/models/*",      # Models tested via integration tests
    "apps/api/migrations/*",       # Migration files don't need coverage
    "apps/api/app/__init__.py",
    "apps/api/app/*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70
